[fix]命名の修正、フォルダ整理、ファイルごとにコード分離
